improvement issue 0000127&0000301:   o Fix Missing Adviser tests   o Improve pwg_high , set to is_the_guest default value for enabled_high
git-svn-id: http://piwigo.org/svn/trunk@1106 68402e56-0260-453c-a942-63ccdbb3a9ee

+++ admin/cat_list.php
@@ -78,8 +78,8 @@ $navigation.= '</a>';
 // +-----------------------------------------------------------------------+
 // |                    virtual categories management                      |
 // +-----------------------------------------------------------------------+
-// request to delete a virtual category
-if (isset($_GET['delete']) and is_numeric($_GET['delete']))
+// request to delete a virtual category / not for an adviser
+if (isset($_GET['delete']) and is_numeric($_GET['delete']) and !is_adviser())
 {
   delete_categories(array($_GET['delete']));
   array_push($page['infos'], $lang['cat_virtual_deleted']);
